+++ dwarf.go
@@ -1,21 +1,7 @@
-// This file is part of GoRE.
+// Copyright (C) 2019-2026 GoRE Authors.
 //
-// Copyright (C) 2019-2024 GoRE Authors
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-//go:build !gore_bsd
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 
 package gore
