+++ slow_test.go
@@ -309,6 +309,26 @@ func TestSourceInfo(t *testing.T) {
 	})
 }
 
+func TestDwarfString(t *testing.T) {
+	noStrip := false
+	getMatrix(t, nil, &noStrip, "dwarfString", func(t *testing.T, exe string) {
+		r := require.New(t)
+
+		f, err := Open(exe)
+		r.NoError(err)
+		r.NotNil(f)
+		defer f.Close()
+
+		gover, ok := getBuildVersionFromDwarf(f.fh)
+		r.True(ok)
+		r.Equal(gover, runtime.Version())
+
+		goroot, ok := getGoRootFromDwarf(f.fh)
+		r.True(ok)
+		r.Equal(goroot, runtime.GOROOT())
+	})
+}
+
 type buildResult struct {
 	exe   string
 	dir   string
